chore: add unit tests for Command params
