+++ src/cli/run.rs
@@ -1,5 +1,6 @@
 use crate::app_config::get_app_config;
-use crate::config::{Command, Config};
+use crate::command::Command;
+use crate::config::Config;
 use crate::env::get_env;
 use crate::ephenv::get_ephenvs;
 use crate::http::handle_request;
