+++ src/constants.rs
@@ -1,0 +1,1 @@
+pub const CONFIG_DIR: &str = ".hit";
