+++ src/main.rs
@@ -1,6 +1,8 @@
 mod app_config;
 mod cli;
+mod command;
 mod config;
+mod constants;
 mod env;
 mod ephenv;
 mod http;
